requirements.txt
fastapi==0.95.0
pymongo==4.3.3
langchain==0.0.185
openai==0.27.4
python-dotenv==0.21.0
requests==2.28.2
uvicorn==0.22.0
pandas==1.5.3